requirements.txt
pymongo
praw
pandas
kafka-python
pyspark
python-dotenv